refactor: update name of supplemental module
